DRTII-1969 Move summary tests to summary spec
